refactor: move io.py in root of service
